some refactoring, plus added support for a Float subtype, plus addressed some code review issues

+++ core/src/test/java/tech/tablesaw/io/csv/CsvReaderTest.java
@@ -45,7 +45,7 @@ import static tech.tablesaw.api.ColumnType.*;
  */
 public class CsvReaderTest {
 
-    private final ColumnType[] bus_types = {INTEGER, STRING, STRING, DOUBLE, DOUBLE};
+    private final ColumnType[] bus_types = {INTEGER, STRING, STRING, FLOAT, FLOAT};
     private final ColumnType[] bus_types_with_SKIP = {INTEGER, STRING, SKIP, DOUBLE, DOUBLE};
 
     @Test
